Switch to using doubles internally
Prevents us from having to deal with floating point error within
the allowable input/output range.

+++ Source/Control/Evaluator.hpp
@@ -16,7 +16,7 @@
 class Evaluator
 {
     public:
-    static float evaluate(List<EquationElement>*);
+    static double evaluate(List<EquationElement>*);
 };
 
 #endif
